refactor: Rename PlayerUI into VisualElement and move some classes around
