+++ arcade-visuals/build.gradle.kts
@@ -5,6 +5,7 @@ moduleDependencies(project, listOf(
 ))
 
 dependencies {
+    // TODO: Remove
     include(modApi(libs.sgui.get())!!)
     modImplementation(libs.polymer.virtual.entity)
 }
