+++ arcade-visuals/src/main/kotlin/net/casual/arcade/visuals/shapes/segment/ShapeSegment.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.visuals.utils.impl
+package net.casual.arcade.visuals.shapes.segment
 
 import net.minecraft.world.phys.Vec3
